* allow advanced users to quieten the installer warning if they know what they are doing

+++ bumblebee/index.php
@@ -72,7 +72,7 @@ if (! $auth->isLoggedIn()) {
   echo $auth->loginError();
 }
 
-if ($auth->isSystemAdmin() && file_exists("install")) {
+if ($auth->isSystemAdmin() && file_exists("install") && ! $conf->value('error_handling', 'ignore_installer', false)) {
   printf('<div class="error">%s</div>',
       T_('The installer still exists. This is a security risk. Please delete it.'));
 }
